claude: refactor(kaspa): extract PSKT utilities to dym_kas_core
Phase 2A of Kaspa refactoring - move pure Kaspa PSKT utilities from
rust/main to libs/kaspa/lib/core:

- input_sighash_type() and is_valid_sighash_type() for standard sighash
- PopulatedInput type alias and PopulatedInputBuilder struct
- utxo_reference_from_populated_input() for mass calculation
- estimate_mass() for transaction mass estimation

These are pure Kaspa utilities with no Hyperlane dependencies, making
libs/kaspa/lib/core more self-contained for PSKT operations.
