+++ dymension/libs/kaspa/lib/core/Cargo.toml
@@ -8,6 +8,7 @@ hardcode = { path = "../hardcode" }
 hex = { workspace = true }
 kaspa-addresses = { workspace = true }
 kaspa-consensus-core = { workspace = true }
+kaspa-consensus-client = { workspace = true }
 kaspa-grpc-client = { workspace = true }
 kaspa-rpc-core = { workspace = true }
 kaspa-core = { workspace = true }
