+++ rust/main/chains/dymension-kaspa/src/kas_relayer/withdraw/messages.rs
@@ -9,18 +9,16 @@ use crate::kas_relayer::withdraw::sweep::{
 };
 use dym_kas_core::consts::RELAYER_SIG_OP_COUNT;
 use dym_kas_core::escrow::EscrowPublic;
+use dym_kas_core::pskt::PopulatedInput;
 use dym_kas_core::wallet::EasyKaspaWallet;
 use dym_kas_hardcode::tx::{MAX_MASS_MARGIN, SWEEPING_THRESHOLD};
 use eyre::Result;
 use hyperlane_core::HyperlaneMessage;
 use hyperlane_core::U256;
 use hyperlane_cosmos::{native::ModuleQueryClient, CosmosProvider};
-use kaspa_consensus_core::tx::{TransactionInput, TransactionOutpoint, UtxoEntry};
+use kaspa_consensus_core::tx::TransactionOutpoint;
 use kaspa_wallet_pskt::bundle::Bundle;
 use tracing::{error, info};
-
-// (input, entry, optional_redeem_script)
-pub(crate) type PopulatedInput = (TransactionInput, UtxoEntry, Option<Vec<u8>>);
 
 /// Adjusts outputs and messages to fit within available funds from swept inputs
 /// Returns (adjusted_outputs, adjusted_messages)
