+++ rust/main/chains/dymension-kaspa/src/kas_validator/withdraw.rs
@@ -2,10 +2,11 @@
 
 use crate::kas_bridge::payload::MessageIDs;
 use crate::kas_bridge::util;
-use crate::kas_bridge::util::{get_recipient_script_pubkey, is_valid_sighash_type};
+use crate::kas_bridge::util::get_recipient_script_pubkey;
 use crate::kas_bridge::withdraw::{filter_pending_withdrawals, WithdrawFXG};
 use crate::kas_validator::error::ValidationError;
 use dym_kas_core::escrow::*;
+use dym_kas_core::pskt::is_valid_sighash_type;
 use dymension_kaspa_hl_constants::ALLOWED_HL_MESSAGE_VERSION;
 use eyre::Result;
 use hex::ToHex;
